Bayesian Models for Age Estimation with Priors from ModelAgePrior Dataset

This function computes the second part of Combes & Philippe (2017), specifically the age estimation using a Bayesian model. Its behavior is similar to other functions like AgeS_Computation(), with the primary difference being the first parameter.

Usage

Compute_AgeS_D(
  DATA,
  Nb_sample,
  SampleNames,
  ThetaMatrix,
  PalaeodoseObject = NULL,
  StratiConstraints = c(),
  model = NULL,
  Iter = 10000,
  burnin = 4000,
  adapt = 1000,
  t = 5,
  n.chains = 3,
  prior = "Jeffreys",
  PriorAge = rep(c(0.01, 100), Nb_sample),
  jags_method = "rjags",
  autorun = F,
  quiet = F,
  roundingOfValue = 3,
  SavePdf = FALSE,
  OutputFileName = c("MCMCplot", "summary"),
  OutputFilePath = c(""),
  SaveEstimates = FALSE,
  OutputTableName = c("DATA"),
  OutputTablePath = c(""),
  ...
)

Arguments

DATA

(required) list The output of the function create_MeasuresDataFrame(), containing the necessary input data for computation.

Value

See the documentation for AgeS_Computation() for details on the returned output.

See also