Compute Isotonic Regression for different stratigraphic constraints

This function compute the isotonic distorsion of a posterior distrubution obtained by Compute_AgeS_D()

Usage

IsotonicCurve(
  StratiConstraints,
  object,
  level = 0.95,
  method = "SBM",
  graphPath = file.path(tempdir(), "graph.html"),
  interactive
)

Arguments

StratiConstraints

: The stratigraphic matrix, it can be stored in a file or given directly. If there is none, then the model suppose that it is a strict order

object

: the returned object of the age model function Compute_AgeS_D()

level

= 0.95 by default for the level of High Posterior Densities regions