Plot graphs build to vizualise isotonic distorsion of the function IsotonicCurve(). The user needs to note that only one HPD regions / Credible Interval level can be allowed and vizualized.

Plot graphs build to vizualise isotonic distorsion of the function IsotonicCurve(). The user needs to note that only one HPD regions / Credible Interval level can be allowed and vizualized.

Usage

PlotIsotonicCurve(object, level = 0.95)

Arguments

object

list: output of the function IsotonicCurve() when using the prior no_strat.

level

numeric 0.95 by default. One of the computed level given in IsotonicCurve() Notice that only one level can be considered.

Value

** NUMERICAL OUTPUT : A list of type BayLum.list containing the following objects**

ribbon : Ribbon of the Credible Interval and solid orange line representing the Bayes mean estimator;

dag : The DAG constructed from the StratiConstraints matrix with segments length represent IC(level) and gradient color the value of ages for each samples

Iso : Data frame used for generating curve and DAG plots. This data frame is extracted from the Ages element of the list object returned by IsotonicCurve()