Compute Isotonic Regression for different stratigraphic constraints

This function compute the isotonic distorsion of a posterior distrubution obtained by Compute_AgeS_D().
The efficient algorithm considered for the Isotnic Regression (IR) is the ** Sequetial Block Merging (SBM)**.
If the StratiConstraints input is a null matrix then the model suppose a strict order

Usage

IsotonicCurve(
  StratiConstraints,
  object,
  levels = c(0.68, 0.95),
  path = tempdir(),
  interactive
)

Arguments

StratiConstraints

matrix or character : The stratigraphic relation between samples.

object

list: output of the function Compute_AgeS_D() when using the prior no_strat. If the StratiConstraints input is a null matrix then the model suppose a strict order

levels

numeric c(0.95, 0.68) by default for the level of High Posterior Densities (HPD) regions. If the HPD region is composed of more than 1 interval then the model return the Credible Interval at the level indicated.

path

character (Default) path for saving the StratiConstraints's DAG

interactive

logical (Default) Indicating wether we want an interactive html file or a plot image for the DAG structure

Value

** NUMERICAL OUTPUT : A list of type BayLum.list containing the following objects**

Sampling : Samples from the posterior distribution after distorsion by the isotonic regression;

network : The DAG constructed from the StratiConstraints matrix

Ages : data frame containing the Credible interval at 95% and 68% , the bayes mean estimator, the bayes standard deviation estimator, the MAP and sample names.

See also